+++ flag/service/service.go
@@ -3,10 +3,10 @@ package service
 import (
 	"github.com/giantswarm/operatorkit/v7/pkg/flag/service/kubernetes"
 
-	"github.com/giantswarm/cert-operator/v2/flag/service/app"
-	"github.com/giantswarm/cert-operator/v2/flag/service/crd"
-	"github.com/giantswarm/cert-operator/v2/flag/service/resource"
-	"github.com/giantswarm/cert-operator/v2/flag/service/vault"
+	"github.com/giantswarm/cert-operator/v3/flag/service/app"
+	"github.com/giantswarm/cert-operator/v3/flag/service/crd"
+	"github.com/giantswarm/cert-operator/v3/flag/service/resource"
+	"github.com/giantswarm/cert-operator/v3/flag/service/vault"
 )
 
 type Service struct {