+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2022-11-23
+
 ### Added
 
 - Add possibility to run cert-operator as a unique app, reconciling special version '0.0.0'.
@@ -115,7 +117,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/cert-operator/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/giantswarm/cert-operator/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/giantswarm/cert-operator/compare/v2.0.1...v3.0.0
 [2.0.1]: https://github.com/giantswarm/cert-operator/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/giantswarm/cert-operator/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/giantswarm/cert-operator/compare/v1.2.0...v1.3.0